Convert negative margin hooks to plain functions, move to atoms
