+++ lib/atoms/negativeMargin/negativeMargin.ts
@@ -1,0 +1,2 @@
+export { negativeMarginLeft } from './negativeMarginLeft';
+export { negativeMarginTop } from './negativeMarginTop';
