+++ lib/atoms/negativeMargin/negativeMarginLeft.ts
@@ -1,0 +1,16 @@
+import classnames from 'classnames';
+import { RequiredResponsiveValue } from '../../atoms/sprinkles.css';
+import { resolveResponsiveProp } from '../../utils/resolveResponsiveProp';
+import * as styles from './negativeMarginLeft.css';
+
+type NegativeMarginLeft = Extract<
+  Extract<keyof typeof styles.mobile, keyof typeof styles.tablet>,
+  Extract<keyof typeof styles.mobile, keyof typeof styles.desktop>
+>;
+
+export const negativeMarginLeft = (
+  space: RequiredResponsiveValue<NegativeMarginLeft>,
+) =>
+  classnames(
+    resolveResponsiveProp(space, styles.mobile, styles.tablet, styles.desktop),
+  );
