--- lib/hooks/useNegativeMargin/useNegativeMargin.ts
@@ -1,2 +1,0 @@
-export { useNegativeMarginLeft } from './useNegativeMarginLeft';
-export { useNegativeMarginTop } from './useNegativeMarginTop';
