--- lib/hooks/useNegativeMargin/useNegativeMarginTop.ts
@@ -1,21 +1,0 @@
-import classnames from 'classnames';
-import { RequiredResponsiveValue } from '../../atoms/sprinkles.css';
-import { resolveResponsiveProp } from '../../utils/resolveResponsiveProp';
-import * as styles from './useNegativeMarginTop.css';
-
-type NegativeMarginTop = Extract<
-  Extract<keyof typeof styles.mobile, keyof typeof styles.tablet>,
-  Extract<keyof typeof styles.mobile, keyof typeof styles.desktop>
->;
-
-export const useNegativeMarginTop = (
-  space: RequiredResponsiveValue<NegativeMarginTop>,
-) => {
-  const negativeMarginTop = resolveResponsiveProp(
-    space,
-    styles.mobile,
-    styles.tablet,
-    styles.desktop,
-  );
-  return classnames(styles.base, negativeMarginTop);
-};
